+++ packages/react-components-lab/src/components/StaticLegend/types.ts
@@ -1,0 +1,9 @@
+import { LegendBaseProps } from '../LegendBase/types';
+
+export interface StaticLegendItem {
+  color: string;
+  label: string;
+}
+export interface StaticLegendProps extends Omit<LegendBaseProps, 'children'> {
+  items: StaticLegendItem[];
+}