+++ packages/react-components-lab/src/index.ts
@@ -1,8 +1,17 @@
-// Legend
-export { default as Legend } from './components/Legend/Legend';
-export * from './components/Legend/Legend';
-export { default as LegendStyles } from './components/Legend/styles';
-export * from './components/Legend/types';
+// Legend (for use with number & color range)
+export { default as ContinuousLegend } from './components/ContinuousLegend/ContinuousLegend';
+export * from './components/ContinuousLegend/ContinuousLegend';
+export * from './components/ContinuousLegend/types';
+
+// StaticLegend
+export { default as StaticLegend } from './components/StaticLegend/StaticLegend';
+export * from './components/StaticLegend/types';
+export { default as useStaticLegendStyles } from './components/StaticLegend/styles';
+
+// LegendBase
+export { default as LegendBase } from './components/LegendBase/LegendBase';
+export * from './components/LegendBase/types';
+export { default as useLegendBaseStyles } from './components/LegendBase/styles';
 
 // BarLegend
 export { default as BarLegend } from './components/BarLegend/BarLegend';